source/webservices/requirements.txt
flask==1.0.2
flask-sqlalchemy==2.3.2
flask-migrate==2.3.0
flask-script==2.0.6
cryptography==2.4.2